requirements.txt
openpyxl
python-docx
docx2pdf
holidays